Commit de rutina

+++ database/migrations/2024_04_09_141554_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string("name",255);
             $table->double("item_price",12,2);
             $table->float("duration");
-            $table->integer("quantity");
+            $table->float("quantity");
             $table->double("sum",12,2);
             $table->boolean("enabled_flag");
             $table->dateTime("add_time");
